feat: add class for listeners
